fix: use float32 instead of float (=64)

+++ examples/params_cont.py
@@ -17,4 +17,4 @@ param_table = np.array([
     [0.01,    0.0,    np.pi], # pa_l
     [-1.2,   -5,    -1],   # log_gamma
     [0.9,   0.0,    np.pi/2],   # pa_gamma
-]).astype(float)
+]).astype("float32")
